fix(receipt): print fractional exchange quantities as issued

+++ src/VvCash/Models/ReturnReceiptLine.cs
@@ -1,3 +1,7 @@
 namespace VvCash.Models;
 
-public record ReturnReceiptLine(string Name, int Quantity, decimal LineRefund);
+/// <summary>Строка чека возврата и обмена. Quantity — decimal, потому что выданная
+/// сторона обмена может быть дробной (1.4 кг). Возвращаемая сторона честно целая:
+/// ReturnLineVm.ReturnQty — int, серверный ReturnLineRequest.Quantity тоже, так что
+/// decimal покрывает оба случая без выдумывания дробных возвратов.</summary>
+public record ReturnReceiptLine(string Name, decimal Quantity, decimal LineRefund);
